refactor: effect slider
