+++ js/effects/effects.js
@@ -83,4 +83,5 @@ const effects = {
   }
 };
 
+
 export { effects };
