+++ js/effects/slider-control.js
@@ -3,22 +3,11 @@ import { effects } from './effects.js';
 
 const sliderContainer = document.querySelector('.img-upload__effect-level');
 const slider = sliderContainer.querySelector('.effect-level__slider');
-const levelValue = sliderContainer.querySelector('.effect-level__value');
 
 const imgContainer = document.querySelector('.img-upload__preview');
 
+
 let currentEffect = 'none';
-
-
-noUiSlider.create(slider, {
-  range: {
-    min: 0,
-    max: 100
-  },
-  start: 0,
-  connect: 'lower'
-});
-
 
 const getFilterValue = (value) => {
   const styleParts = effects[currentEffect].filter;
@@ -43,17 +32,8 @@ const updateSlider = (effect) => {
 
   toggleSliderVisibility(effect);
 
-  slider.noUiSlider.updateOptions(effects[currentEffect].options);
+  slider.noUiSlider.updateOptions(effects[effect].options);
 };
 
 
-slider.noUiSlider.on('update', () => {
-  const value = slider.noUiSlider.get();
-
-  levelValue.value = value;
-
-  updateEffect(value);
-});
-
-
-export { updateSlider };
+export { updateSlider, updateEffect };
