+++ js/main.js
@@ -2,7 +2,10 @@ import { fillPictures } from './minis.js';
 import './modal/photo-edit.js';
 import './validation/validation.js';
 import './scale-control.js';
-import './effects/level.js';
-import './effects/selection.js';
+import { initEffectSlider } from './effects/slider-init.js';
+import { initEffectSelector } from './effects/selection.js';
 
 fillPictures();
+
+initEffectSlider();
+initEffectSelector();
